% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create.graph}
\alias{create.graph}
\title{creates an adjacency graph from the nodes in provParseR}
\usage{
create.graph()
}
\value{
nothing
}
\description{
prov.parse must be called before this function
}
\examples{
provParseR::prov.parse(system.file("testdata", "prov.json", package = "provGraphR"))
create.graph()
}
