% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get.spine}
\alias{get.spine}
\title{Returns lineage of provided node}
\usage{
get.spine(node.id, forward = F)
}
\arguments{
\item{node.id}{This is a label for a node that the lineage is
being requested for}

\item{forward}{Logical that states whether or not to transpose
the graph allowing the function to search forward or backward
in lineage}
}
\value{
a vector of characters, each will be a different node label
}
\description{
Returns lineage of provided node
}
