% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get.creator}
\alias{get.creator}
\title{Get creators and users of data}
\usage{
get.creator(adj.graph, node.id)
}
\arguments{
\item{adj.graph}{the adjacency matrix}

\item{node.id}{the id of the data node.}
}
\value{
the id of the procedure node that created the specified data node.  
  Returns NULL if there is no node with the given id, the id is not for 
  a data node, or the data node does not have a creator.  The last case
  can occur, for example, if the data node represents an input file.
}
\description{
get.creator finds the node that creates the given data node
}
\details{
A data node can represent a variable, a file, a plot, or a warning or error.  The creator
of the data node will be a procedure node representing the statment
that assigned the variable, wrote to the file, created the plot, or 
resulted in the error or warning.
}
