% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create.graph}
\alias{create.graph}
\title{Create provenance graph}
\usage{
create.graph(prov.input = NULL, isFile = T)
}
\arguments{
\item{prov.input}{This is either a file name or a string contain provenance
in PROV-JSON format.  PROV-JSON can be generated by RDataTracker
or provR.  If this parameter is
missing, and prov.parse has been called previously, that parsed provenance
will be used.}

\item{isFile}{A logical that determines whether or not the prov.input parameter
contains a file name or contains the provenance in a string.  This parameter is
ignored if prov.input is missing.}
}
\value{
create.graph returns a matrix representation of the graph.  There is a row and a column for each
  data and procedure node in the graph.  The values in the matrix are either 1 or 0.
  A 1 indicates that there is an edge for the column node to the row node.  create.graph returns
  NULL if there is no provenance available.
}
\description{
create.graph creates an adjacency graph containing the procedure and
data nodes in saved provenance.
}
\examples{
create.graph(system.file("testdata", "prov.json", package = "provGraphR"))
}
