% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get.lineage}
\alias{get.lineage}
\title{Calculate lineage of a node}
\usage{
get.lineage(adj.graph, node.id, forward = F)
}
\arguments{
\item{adj.graph}{An adjacency graph to get the lineage from.  The
graph can be created by a call to create.graph.}

\item{node.id}{This is a string id for a node that the lineage is
being requested for}

\item{forward}{Logical that states whether the search is going forward
through the graph from the provided node, or backwards.}
}
\value{
get.lineage returns the forward or backward lineage of the specified node.  The lineage
  is represented as a vector of character, with each being a different node label.
}
\description{
get.lineage returns the lineage of the provided node.  Forward lineage identifies
everything that depends on the provided node.  Backward lineage identifies
everything the provided node depends on.
}
\details{
Most commonly, the node passed in is a data node representing either a variable,
a file, or a plot.  Forward lineage reports everything computed from that variable
or file.  Backward lineage reports everything that contributed to the variable's 
value, the contents of an output file or plot.
}
\examples{
adj.graph <- create.graph(system.file("testdata", "prov.json", package = "provGraphR"))
get.lineage (adj.graph, "d33")

}
\seealso{
\code{\link{create.graph}}
}
